feat/fix: make on-change's default "before" behaviour optional
on-change's default behaviour is to fire the callback (onChange) *before* the value has actually been set/updated in the observed object. This is natural Proxy API behaviour (since the callback is fired from the setters before returning from them). But this behaviour is unexpected when we have to actually update data in storage - where we must wait for the update to take place before writing, otherwise we're just writing the previous data.